add getHistoryInterval

+++ src/main/kotlin/HistoryTable.kt
@@ -2,13 +2,14 @@ package me.mixo.library
 
 import org.ktorm.schema.Table
 import org.ktorm.schema.int
+import org.ktorm.schema.long
 import org.ktorm.schema.text
 
 object HistoryTable : Table<Nothing>("history") {
     val id = int("_id").primaryKey()
     val title = text("title")
-    val created = int("created")
-    val last_date = int("date")
+    val created = long("created")
+    val last_date = long("date")
     val visits = int("visits")
     val url = text("url")
     val user_entered = int("user_entered")
@@ -18,8 +19,8 @@ object HistoryTable : Table<Nothing>("history") {
 data class HistoryData(
     val id:Int,
     val title:String,
-    val created:Int,
-    val last_date:Int,
+    val created:Long,
+    val last_date:Long,
     val url :String,
     val visits:Int,
     val user_entered:Int,
